Revert changing the entity version to be signed

+++ packages/graph/migrations/postgres/migration/1656417312397_initial.ts
@@ -56,7 +56,7 @@ export async function up(pgm: MigrationBuilder): Promise<void> {
         onDelete: "CASCADE",
       },
       version: {
-        type: "INT",
+        type: "BIGINT",
         notNull: true,
       },
       version_id: {
